requirements.txt
chalice>=1.0
psycopg2-binary==2.9.6
idna~=3.4
chardet~=5.1.0
certifi==2023.5.7
postgres~=4.0
boto3==1.17.105
urllib3==1.26.7
requests==2.26.0
pyjwt==2.3.0
cryptography==3.4.8